pyproject.toml
[tool.poetry]
name = "spotter"
version = "0.1.0"
description = ""
authors = ["Manish Tekam <manish.tekam.9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
scikit-learn = "^1.5.2"
numpy = "^2.1.1"
celery = "^5.4.0"
rabbitmq = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
